calendar_assistant/requirements.txt
google-api-python-client>=2.131.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
python-telegram-bot>=21.0
openai>=1.16.0
python-dateutil>=2.9.0
pytz>=2024.1
